Add jitter to raytracer

+++ src/raytracer/basic.rs
@@ -17,7 +17,7 @@ impl<'r, DS: DataStructure<'r>, S: Shader<'r, DS>> RayTracer<'r, DS, S> for Basi
 
         for x in 0..camera.width {
             for y in 0..camera.height {
-                let ray = camera.generate_ray(x, y);
+                let ray = camera.generate_ray(x as f64, y as f64);
 
                 if let Some(intersection) = datastructure.intersects(&ray) {
                     output.set_at(x, y, shader.shade(&intersection,datastructure).into());
